requirements.txt
# Core dependencies
aiohttp>=3.9.0

# CLI and UI
rich>=13.7.0

# Configuration
pyyaml>=6.0.1

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Type checking (optional)
mypy>=1.7.0
types-PyYAML>=6.0.12
